Optimized dictionary creation to use a pre-coded list of hashmap inserts. Created dict_code_noplurals.rs and dict_code_all.rs and renamed dicionary.rs to create_dictionary.rs Total execution time: 40-60ms
Todo: try to load the words as a static array of vectors.  Then either
recode the functions to take an array or change the hashmap to reference a
vector and do the inserts for each length with a reference to the vector
in the array index of the same length.